Extract code for S3 download
